server/requirements.txt
django==2.1.7
djangorestframework
django-cors-headers
gunicorn==19.7.1